Move debug commands under debug category
