+++ src/cli/commands/debug/bisect.ts
@@ -1,4 +1,4 @@
-import type { MetricStore } from "../storage/types.js";
+import type { MetricStore } from "../../storage/types.js";
 
 export interface BisectResult {
   lastGoodCommit: string;
