+++ src/cli/commands/debug/confirm.ts
